requirements.txt
adal==1.2.7
# Editable Git install with no remote (autopipes==0.0.0)
-e /Users/shaunryan/autopipes
Babel==2.15.0
backports.tarfile==1.2.0
black==24.4.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.5.3
cryptography==42.0.8
docutils==0.21.2
flake8==7.0.0
ghp-import==2.1.0
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
Jinja2==3.1.4
keyring==25.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.26
mkdocs-material-extensions==1.3.1
more-itertools==10.2.0
mypy-extensions==1.0.0
nh3==0.2.17
packaging==24.0
paginate==0.5.6
pathspec==0.12.1
pkginfo==1.11.1
platformdirs==4.2.2
pluggy==1.5.0
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pymdown-extensions==10.8.1
pytest==8.2.2
pytest-cov==5.0.0
pytest-env==1.1.3
pytest-mock==3.14.0
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyyaml_env_tag==0.1
readme_renderer==43.0
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
six==1.16.0
twine==5.1.0
urllib3==2.2.1
watchdog==4.0.1
zipp==3.19.2
